+++ test/goldens/wave1/models/models.dart
@@ -3,3 +3,4 @@
 
 export 'sessions.dart';
 export 'songs.dart';
+export 'session_players.dart';